Add profile rework including media profiles
The structure of profiles has been altered to support
multiple types of profiles.
