+++ package.json
@@ -15,15 +15,17 @@
     "graphql": "^14.5.8",
     "js-sha256": "^0.9.0",
     "lodash.flowright": "^3.5.0",
-    "mdbreact": "mdbreact-4.25.1.tgz",
+    "mdbreact": "mdbreact-4.25.3.tgz",
     "moment": "^2.24.0",
     "node-sass": "^4.13.1",
     "obelisk.js": "^1.2.2",
     "react": "^16.8.6",
     "react-apollo": "^3.1.3",
     "react-dom": "^16.8.6",
+    "react-instagram-embed": "^1.5.0",
     "react-router-dom": "^5.0.1",
     "react-scripts": "^3.3.0",
+    "react-simple-maps": "^2.0.0",
     "react-text-loop": "^2.3.0",
     "react-typed": "^1.2.0",
     "react-very-simple-oauth": "^1.1.0",
