+++ src/App.js
@@ -121,7 +121,6 @@ const UPDATE_CACHE = gql`
   }
 `;
 class App extends React.Component {
-
   state = {
     logged: false,
     user: undefined,
@@ -215,6 +214,66 @@ class App extends React.Component {
         let platformData = JSON.parse(data.profile.platformData);
         let sources = JSON.parse(data.profile.sources);
         let cache = {};
+        /**
+         * ################
+         * DUMMY DATA
+         * Remove and replace with live data when ready
+         * ################
+         */
+        // Change this to change from software to media
+        let enableMediaEngineer = true;
+        // Injecting platformData
+        if(enableMediaEngineer){
+          // Set type to media to distinguish
+          platformData.user.type = "media";
+          // Set media engineer platforms
+          platformData.user.platforms = {
+            instagram: {
+              url: "https://www.instagram.com/aichnerchristian/"
+            },
+            facebook: {
+              url: "https://www.facebook.com/aichner.christian"
+            },
+            portfolio: {
+              url: "https://www.aichner-christia.com/portfolio"
+            }
+          }
+          // Portfolio map
+          platformData.user.mapData = [
+            { name: "1", coordinates: [12.8506, 44.6086] },
+            { name: "2", coordinates: [13.8496928, 46.6114363 ] },
+            { name: "3", coordinates: [11.489387, 48.783450 ] }
+          ]
+          // Settings
+          platformData.user.settings = {
+            showMap: true,
+            showInstagramFeed: true,
+            instagramHideCaption: true,
+          }
+          // Skills (like languages for programmers)
+          platformData.user.skills = [
+            {name: "Photography", color: "#563d7c", size: 54, share: 10},
+            {name: "Video", color: "#263d1c", size: 54, share: 20},
+            {name: "Web", color: "#763d2c", size: 54, share: 70},
+          ]
+          // Instagram posts
+          platformData.user.instagram = [
+            {url: "https://www.instagram.com/p/B9cOSWMJbXD/"},
+            {url: "https://www.instagram.com/p/B9TWGNaglUz/"}
+          ]
+        } else {
+          // Add needed variables software engineer
+          platformData.user.type = "software";
+        }
+        // Add needed variables for both software- and media engineer
+        platformData.user.first_name = "Max";
+        platformData.user.last_name = "Mustermann";
+        console.log("PD",platformData);
+        /**
+         * ################
+         * DUMMY DATA END
+         * ################
+         */
         this.setState({
           fetchedUser: {
             platformData: platformData,
@@ -227,7 +286,7 @@ class App extends React.Component {
             }
           },
         });
-        intel
+        /*intel
         .fill(sources)
         .then(async () => {
           intel.calendar();
@@ -262,7 +321,7 @@ class App extends React.Component {
             platformData
           }
           });
-        });
+        });*/
       } else {
         this.setState({
           fetchedUser: false,
