+++ src/components/organisms/profiles/index.js
@@ -1,0 +1,12 @@
+import SoftwareEngineer from "./SoftwareEngineer";
+import MediaEngineer from "./MediaEngineer";
+
+export {
+  SoftwareEngineer,
+  MediaEngineer,
+}
+
+/**
+ * SPDX-License-Identifier: (EUPL-1.2)
+ * Copyright © 2019 Werbeagentur Christian Aichner
+ */
